comit
This Postal Code Search project is a demo application developed entirely with Node.js, designed to reinforce and practice my existing knowledge.
If the entered postal code does not exist in the database, the application fetches the data from an external API, saves it to the database, and then returns the JSON response to the user.